created dto for miner stats with slight restructuring for gpu share stats
