+++ app/main/model/gpu.py
@@ -1,4 +1,3 @@
-import enum
 from datetime import datetime
 from sqlalchemy.orm import relationship
 from sqlalchemy.dialects.mssql import UNIQUEIDENTIFIER
@@ -10,7 +9,6 @@ from sqlalchemy import (
     Float,
     ForeignKeyConstraint,
     UniqueConstraint,
-    Enum
 )
 from . import Base
 
@@ -57,29 +55,3 @@ class Health(Base):
     def as_dict(self):
         return {c.name: getattr(self, c.name) for c in self.__table__.columns}
 
-
-class ShareType(enum.Enum):
-    valid = "valid share"
-    invalid = "invalid share"
-    stale = "stale share (reported by pool)"
-
-
-class Share(Base):
-    __tablename__ = "share"
-    __table_args__ = (
-        ForeignKeyConstraint(
-            ["miner_id", "gpu_no"], ["gpu.miner_id", "gpu.gpu_no"],
-            ondelete="CASCADE"
-        ),
-    )
-
-    time = Column(DateTime, default=datetime.now(), primary_key=True)
-
-    # don't need backpopulates because we only ever insert into this table
-    miner_id = Column(UNIQUEIDENTIFIER, primary_key=True)
-    gpu_no = Column(Integer, primary_key=True)
-
-    type = Column(Enum(ShareType), nullable=False)
-
-    def as_dict(self):
-        return {c.name: getattr(self, c.name) for c in self.__table__.columns}
